Update header logo router, needs styilng fix
