+++ src/styles/globalStyles.js
@@ -35,3 +35,32 @@ export const StyledInnerBox = styled(Box)({
   justifyContent: { xs: "center", sm: "flex-start" },
   gap: 1,
 });
+
+export const StyledTypography = styled(Typography)({
+  padding: "1rem",
+  fontFamily: "roboto",
+  textShadow: "0 0 10px white",
+  fontWeight: "800",
+  color: "white",
+});
+
+export function Heading(props) {
+  return (
+    <StyledTypography
+      sx={{
+        padding: "1rem",
+        textShadow: "0 0 10px white",
+        fontWeight: "1000",
+        fontSize: {
+          xs: "1rem",
+          sm: "1.5rem",
+          md: "1.8rem",
+          lg: "2rem",
+        },
+        color: "white",
+      }}
+    >
+      {props.children}
+    </StyledTypography>
+  );
+}
